backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pydantic-settings==2.5.0
google-generativeai==0.8.3
edge-tts==6.1.12
genanki==0.13.0
python-multipart==0.0.12
aiofiles==24.1.0
fpdf2==2.7.9
faster-whisper==1.0.3
feedparser==6.0.11
